+++ src/main/java/com/example/jpaschedule/config/interceptor/SessionInterceptor.java
@@ -1,6 +1,6 @@
-package com.example.jpaschedule.filter.interceptor;
+package com.example.jpaschedule.config.interceptor;
 
-import com.example.jpaschedule.filter.context.MemberContext;
+import com.example.jpaschedule.config.context.MemberContext;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
