+++ src/main/java/com/example/jpaschedule/domain/service/MemberService.java
@@ -4,8 +4,8 @@ import com.example.jpaschedule.domain.dto.request.MemberRequestDto;
 import com.example.jpaschedule.domain.dto.response.MemberResponseDto;
 import com.example.jpaschedule.domain.entity.Member;
 import com.example.jpaschedule.domain.repository.MemberRepository;
-import com.example.jpaschedule.filter.context.MemberContext;
-import com.example.jpaschedule.util.EmptyTool;
+import com.example.jpaschedule.config.context.MemberContext;
+import com.example.jpaschedule.common.util.EmptyTool;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
