+++ src/main/java/com/example/jpaschedule/domain/service/ScheduleService.java
@@ -5,8 +5,8 @@ import com.example.jpaschedule.domain.dto.response.ScheduleResponseDto;
 import com.example.jpaschedule.domain.entity.Member;
 import com.example.jpaschedule.domain.entity.Schedule;
 import com.example.jpaschedule.domain.repository.ScheduleRepository;
-import com.example.jpaschedule.filter.context.MemberContext;
-import com.example.jpaschedule.util.EmptyTool;
+import com.example.jpaschedule.config.context.MemberContext;
+import com.example.jpaschedule.common.util.EmptyTool;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
